+++ CHANGELOG.md
@@ -6,7 +6,11 @@
 
 ## 10.6.1
 * Changed Enhanced SmartSelfie™ viewmodels to call `onResult` when the user clicks cancel instead of when there is an error    
-* Added messages from SmartSelfie™ errors will show correctly if present and will show default failure message if not  present  
+* Added messages from SmartSelfie™ errors will show correctly if present and will show default
+  failure message if not present
+* Changed File save path from cacheDir to fileDir (and maintain compatibility till next major
+  breaking release)
+* Updated the preview to throw SmileIDException and not crash if the file is null
 
 ## 10.6.0
 * Changes the `allow_new_enroll` flag to be a real boolean instead of a string for prepUpload